refactor(desktop): consolidate duplicate HiringBanner into renderer/components
