+++ apps/desktop/src/renderer/components/HiringBanner/index.ts
