--- apps/desktop/src/renderer/routes/_authenticated/_dashboard/components/DashboardSidebar/components/HiringBanner/index.ts
@@ -1,1 +1,0 @@
-export { HiringBanner } from "./HiringBanner";
