+++ apps/desktop/src/renderer/screens/main/components/WorkspaceSidebar/WorkspaceSidebar.tsx
@@ -1,7 +1,7 @@
 import { useCallback, useEffect, useMemo } from "react";
+import { HiringBanner } from "renderer/components/HiringBanner";
 import { useWorkspaceShortcuts } from "renderer/hooks/useWorkspaceShortcuts";
 import { useWorkspaceSelectionStore } from "renderer/stores/workspace-selection";
-import { HiringBanner } from "./HiringBanner";
 import { MultiDragPreview } from "./MultiDragPreview";
 import { PortsList } from "./PortsList";
 import { ProjectSection } from "./ProjectSection";
@@ -116,7 +116,7 @@ export function WorkspaceSidebar({
 				projectName={activeProjectName}
 			/>
 
-			<HiringBanner isCollapsed={isCollapsed} />
+			<HiringBanner surface="v1" isCollapsed={isCollapsed} />
 
 			<WorkspaceSidebarFooter isCollapsed={isCollapsed} />
 			<MultiDragPreview />
